+++ examples/api/lib/material/selection_area/selection_area.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// Flutter code sample for SelectionArea
+// Flutter code sample for SelectionArea.
 
 import 'package:flutter/material.dart';
